+++ purdy/builder.py
@@ -1,4 +1,5 @@
-from typing import Iterable, Iterator, Union, List, Callable, Optional, Any
+from typing import (Iterable, Iterator, Union, List, Callable, Optional, Any,
+    Tuple)
 from purdy.actions import (AppendTypewriter, Append, Suffix, Wait, Clear,
     Insert, Replace, Remove, Shell, InsertTypewriter, SuffixTypewriter,
     Highlight, HighlightChain, Fold, StopMovie, Transition, Sleep, RunFunction,
@@ -304,10 +305,12 @@ class ActionsBuilder(Iterable):
         return self._add_action(Transition(self.__code_box,
                 self._create_code(filename, text, code), code_box_to_copy))
 
-    def sleep(self, time: Union[float, int]) -> "ActionsBuilder":
+    def sleep(self, time: Union[float, int, 
+            Tuple[Union[float, int], Union[float, int]]) -> "ActionsBuilder":
         """Adds an :class:`purdy.actions.Sleep` action
 
-        :param time: amount of time to sleep given in seconds
+        :param time: amount of time to sleep given in seconds, or a tuple
+                     specifying lower and upper bound of random time in seconds 
 
         :return: self
         """